electrical panel in own file, smart fridge good to go
