+++ tgui/packages/tgui/styles/interfaces/SmartFridge.scss
@@ -1,5 +1,3 @@
-$electrical-width: 300px;
-
 .SmartFridge {
   .ItemIconCell {
     width: 32px;
@@ -32,70 +30,4 @@ $electrical-width: 300px;
       margin-bottom: -1px;
     }
   }
-
-  .ElectricalAccessPanel {
-    border-bottom: 2px solid #606060;
-    border-left: 2px solid #606060;
-    border-right: 2px solid grey;
-    border-top: 2px solid grey;
-    width: $electrical-width;
-    display: flex;
-    flex-direction: row;
-    overflow-x: hidden;
-    overflow-y: hidden;
-
-    .PanelClosed {
-      width: $electrical-width;
-      height: 100%;
-    }
-    .PanelOpen {
-      width: $electrical-width;
-      height: 100%;
-      padding: 6px;
-    }
-    .ElectricalSafetySign {
-      height: 100%;
-    }
-    span {
-      display: flex;
-      text-align: center;
-    }
-    .WarningIcon {
-      border: 2px solid black;
-      padding: 3px;
-      font-weight: bold;
-      width: 200px;
-      height: 100%;
-    }
-    .OpenSafetySign {
-      border-top: 1px solid black;
-      margin-bottom: -6px;
-      margin-left: -6px;
-      margin-right: -6px;
-    }
-    .WireLabel {
-      overflow: auto;
-      word-wrap: break-word;
-      width: 100px;
-    }
-    .WirePanel {
-      display: inline-table;
-      margin-bottom: 6px;
-      span {
-        border: 2px solid black;
-      }
-    }
-    .WireControl {
-      height: 50px;
-    }
-    .WireSketch {
-      border: 1px solid white;
-      height: 100%;
-      width: 2px;
-    }
-  }
-}
-
-.Test {
-  border: 4px solid red;
 }
