+++ tgui/packages/tgui/styles/main.scss
@@ -65,6 +65,7 @@
 @include meta.load-css('./interfaces/Techweb.scss');
 @include meta.load-css('./interfaces/TechNode.scss');
 @include meta.load-css('./interfaces/VendingSorted.scss');
+@include meta.load-css('./interfaces/common/ElectricalPanel.scss');
 
 // Layouts
 @include meta.load-css('./layouts/Layout.scss');
